requirements.txt
alembic==1.7.5
astroid==2.9.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
coverage==6.2
coveralls==3.3.1
docopt==0.6.2
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.6.0
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
mccabe==0.6.1
platformdirs==2.4.0
pylint==2.12.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
PyMySQL==1.0.2
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
SQLAlchemy-Utils==0.37.9
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
